Fix #8
Catch IO exceptions that are a result of Windows system error code 32 "The process cannot access the file because it is being used by another process." as this is seems to be the expected behavior for system disks.

Online status will be set to "Unknown" in this event. Rethrow IO exceptions that are not a result of error 32.